apps/bot/pyproject.toml
[project]
name = "usipipo-telegram-bot"
version = "0.15.1"
description = "Bot de Telegram para interacción con usuarios del ecosistema uSipipo"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [{name = "uSipipo Team", email = "usipipo@gmail.com"}]

dependencies = [
    "usipipo-commons>=0.15.0",
    "python-telegram-bot>=21.0",
    "pydantic-settings>=2.0.0",
    "redis>=7.3.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
